+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.3] - 2025-03-05
 - improved topdata_report table structure
 - Added: command to check for crashed jobs and mark them in the report table
+- Added: option to delete reports with no PID using --delete-no-pid flag
 
 
 ## [1.1.0] - 2025-02-27